refactor(#30): move /infra to /config
