+++ backend/zerowaste/src/main/java/com/zerowaste/zerowaste/config/auth/SpringSecurityConfiguration.java
@@ -1,4 +1,4 @@
-package com.zerowaste.zerowaste.infra.auth;
+package com.zerowaste.zerowaste.config.auth;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
