+++ backend/zerowaste/src/main/java/com/zerowaste/zerowaste/repositories/interfaces/UsersRepository.java
@@ -1,8 +1,8 @@
-package com.zerowaste.zerowaste.application.interfaces;
+package com.zerowaste.zerowaste.repositories.interfaces;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.zerowaste.zerowaste.domain.models.user.User;
+import com.zerowaste.zerowaste.models.user.User;
 
 public interface UsersRepository extends JpaRepository<User, Long> {
     public User findByEmail(String email);
