+++ backend/zerowaste/src/main/java/com/zerowaste/zerowaste/services/users/RegisterUserService.java
@@ -5,9 +5,9 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import com.zerowaste.zerowaste.application.interfaces.UsersRepository;
-import com.zerowaste.zerowaste.domain.models.user.User;
-import com.zerowaste.zerowaste.domain.models.user.UserRole;
+import com.zerowaste.zerowaste.models.user.User;
+import com.zerowaste.zerowaste.models.user.UserRole;
+import com.zerowaste.zerowaste.repositories.interfaces.UsersRepository;
 import com.zerowaste.zerowaste.services.users.exceptions.UserWithSameEmailAlreadyExistsException;
 
 @Service
